Name on navbar added
Still as some problems
